requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
bcrypt==4.1.2
blinker==1.8.1
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
exceptiongroup==1.2.1
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
nodeenv==1.8.0
prisma==0.13.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dotenv==1.0.1
sniffio==1.3.1
StrEnum==0.4.15
tomlkit==0.12.4
typing_extensions==4.11.0
waitress==3.0.0
Werkzeug==3.0.2
